+++ example/brick_breaker/index.ts
@@ -269,14 +269,14 @@ class SimpleScene extends Scene {
         player.Add(new Transform(new Vector(0, -40), new Vector(30, 5)));
         player.Add(new Sprite(new Color(Math.random(), Math.random(), Math.random(), 1), 0));
         player.Add(new Motion(new Vector(0, 0), new Vector(0, 0), 0));
-        player.Add(new Collider(Polygon.Rectangle(1, 1)))
+        player.Add(new Collider(Polygon.RectangleByDimensions(1, 1)))
         player.Add(new Player());
 
         const ball = new Entity(this.messageBus);
         ball.Add(new Transform(new Vector(0, -35), new Vector(4, 4)));
         ball.Add(new Sprite(new Color(Math.random(), Math.random(), Math.random(), 1), 0));
         ball.Add(new Motion(new Vector(0, 0), new Vector(0, 0), 0));
-        ball.Add(new Collider(Polygon.Rectangle(1, 1)))
+        ball.Add(new Collider(Polygon.RectangleByDimensions(1, 1)))
         ball.Add(new Ball(false));
 
 
@@ -288,7 +288,7 @@ class SimpleScene extends Scene {
                 const brick = new Entity(this.messageBus);
                 brick.Add(new Transform(new Vector((-brickColumns * 22) / 2 + 22 / 2 + i * 22, 40 - j * 10), new Vector(20, 5)));
                 brick.Add(new Sprite(new Color(Math.random(), Math.random(), Math.random(), 1), 0));
-                brick.Add(new Collider(Polygon.Rectangle(1, 1)))
+                brick.Add(new Collider(Polygon.RectangleByDimensions(1, 1)))
                 brick.Add(new Brick());
             }
         }